Save subset of Typing_env in TAST EnvAnnotation
Summary:
Currently, the TAST defines a type `TAST.program`, which would seem to be the type of a TAST representing an entire file. This isn't actually the case, though--in order to make use of the TAST, it is necessary to use the "tast_tenv" type returned by `Typing.nast_to_tast_tenv` (i.e., a `(Tast.def * Typing_env.env) list`), since type variable substitutions derived during inference are only contained in the typing environment. Additionally, further substitutions are derived during the checking of method definitions, but are not represented in the returned environment at all.

To fix this, this stack adds an "Env" annotation to the annotated AST functor, so that the Typed AST can store typing environment information for subtrees which require it (currently, definitions of classes, methods, functions, and consts).

This diff stores the tenv and subst maps from Typing_env in the annotation for toplevel definitions and methods, extends Aast_mapper to allow consumers to make use of this annotation, and updates hh_single_type_check's TAST mappers to do so.

Reviewed By: andrewjkennedy

Differential Revision: D6255757

fbshipit-source-id: da9dd69d9f548a57b96800efe03ea5a2d9793bfb