Stop returning full typing env with toplevel defs
Summary:
Currently, the TAST defines a type `TAST.program`, which would seem to be the type of a TAST representing an entire file. This isn't actually the case, though--in order to make use of the TAST, it is necessary to use the "tast_tenv" type returned by `Typing.nast_to_tast_tenv` (i.e., a `(Tast.def * Typing_env.env) list`), since type variable substitutions derived during inference are only contained in the typing environment. Additionally, further substitutions are derived during the checking of method definitions, but are not represented in the returned environment at all.

To fix this, this stack adds an "Env" annotation to the annotated AST functor, so that the Typed AST can store typing environment information for subtrees which require it (currently, definitions of classes, methods, functions, and consts).

This diff removes the "tast_tenv" type, replacing instances of `(Tast.def * Typing_env.env) list` with `Tast.program`. Since the TAST's EnvAnnotations now contain the mappings necessary to reconstruct types, it is no longer necessary to return an entire typing environment with each toplevel definition.

Reviewed By: andrewjkennedy

Differential Revision: D6255787

fbshipit-source-id: fc0bcf4d8ea3395c6e814804116e6ea4ac721303