Move stream context http header processing
Commit eec9da990e7344af985735ec53d73325579931eb appended the http header from
the stream context to the curl list of headers. This was done to allow a
SoapClient to be constructed with a stream context that contained http headers.
It created issues #7684 and #8046.

HttpClient::request() has a requestHeaders argument that should be
used instead. Because there are three places that use HttpClient with a
StreamContext object, this commit adds HttpClient.streamContextHttpHeader()
and moves all processing of stream context http headers there.

Places that use HttpClient.setStreamContextOptions()
* SoapClient::__construct(), -> sdl.cpp load_wsdl_ex()
* SoapClient::__dorequest()
* UrlFile->open()

Closes #5553
Closes #7684
Closes #8046
Closes #8047