Reject use of using variables outside using block
Summary:
The implementation of `using` unsets using variables at the end of the using block.
We want to reject uses outside the using block statically. This is something unprecedented
in Hack, where variables don't currently go "out of scope", but it seems possible to achieve
this in the naming phase by removing the locals from the environment.

Reviewed By: dlreeves, KendallHopkins

Differential Revision: D6358353

fbshipit-source-id: 4f1316de7ca42b9b9aa90df177eb5bcab62381ed