Correct retry behavior for tiny states
Summary:
This diff makes it so that turning off the tiny saved states experiment is easy.

It deletes a mini saved state when we expect to be using a tiny saved state (and vice versa).
It retries loading a saved state target if we fail because we needed to clear out the other kind of saved state (which we now do).

Reviewed By: jamesjwu

Differential Revision: D6365726

fbshipit-source-id: f0b1d1d7482e4315c88d6c71a99af7c67adc93ec