typechecking for XHP spread operator
Summary:
This diff adds typechecker support to the XHP spread operator. The typing rules behave as though all of the declared attributes on XHP class(es) from the `{...$expr}` were given as `attr := Tattr` in order exactly where the spread appears in the attribute list. The actual copying behavior of spreads happens at runtime, but it will preserve the semantics as the typechecker sees them; this is how all XHP features currently work (runtime-defined code through a special constructor)

In the interest of shipping the feature and at least replacing some unwanted implicit copying/spread behavior in www, there are two small soundness issues that allow some attribute type mismatches to go by unnoticed. The code and unit tests document where the behavior occurs; one issue is not enforcing final or some `<<__ConsistentAttributes>>` marker for XHP classes and the other is unsoundly treating `$this` as `Env.get_self` even when the class is not final.

This diff is a no-op for the existing `attr={value}` way of building XHP attributes; the only side-effect is that error messages now call the attribute `:` because that is the actual class member name from Hack's point of view and, since the typing rules pull all the XHP attribute members from a class, they have `:` built in where we added it before. If this bothers anyone, I can slice the first character off for the error message, but I kind of like it having the `:`.

Reviewed By: dlreeves

Differential Revision: D6349556

fbshipit-source-id: cff55c362eb90152c45430d889915e2a75fd2e10