Move type-at-pos integration test position to paren instead of name
Summary:
This test was inadvertently broken by rFBSfb59b7ec9d0d282dd3c91d31f8aa8aef8a1f848e
which changed to using the TAST for ServerInferType.

Checking the type of the name returns the function type instead of the expression's
type. This updates the position of the test to ask for the type of the open
parenthesis, instead of the name to unbreak this test.

Differential Revision: D6381407

fbshipit-source-id: 983fafe2072e3bea6b27f48cad499b71d68a1c8a