Don't use propLvalAtOffset in variable unserializer.
Summary: I used propLvalAtOffset by mistake in the variable unserializer cold path. It is meant to be used only with properties that are statically known to be declared and mutable, so switch to the slightly more expensive getPropLval instead; it should not matter to perf because this is the cold path. Clarify comments for propLvalAtOffset and propRvalAtOffset to indicate that the property access must be known to be safe statically. Also fix whitespace nits in the definitions of these methods that mwang noted on an earlier diff.

Reviewed By: mxw

Differential Revision: D6386363

fbshipit-source-id: e2311bcd82c78c7a7ff5a4207b02b1fe87fbf294