Emit syntax error fatals with position
Summary:
Update `hh_single_compile` so that it emits position information for parser Fatals.

Only one new test passes. But in most instances, HackC emits *better* error information than HHVM: the position is more accurate (e.g. on the `foreach` loop not on the end), and the message more informative.

Differential Revision: D6384986

fbshipit-source-id: d5e0fa34a0f21f39eb3dcfbf3a5a72f4c60d91c9