Fix bug in LdClsCctx
Summary: We use LdClsCctx to load the class from pre-live act-recs in inlined regions. This logic is seldom used (often the class is constant or already in register), so the bug has been around for sometime now. In this case we refine the type returned by LdCLsCctx to a subclass of the current context, and then in the inliner we perform an AssertType based on the known context type, which in the case of the crashes I've found caused us to mark the entire function as unreachable.

Reviewed By: swtaarrs

Differential Revision: D6394938

fbshipit-source-id: b30eccae13b88601764af48ea8cfcb80a1798a46