Set source location when constructing pseudomain from HHAS
Summary: The (1,1), say, in a .main (1,1) declaration was not actually used to set the source location. Instead, the assembler relied on there being a .srcloc directive in the body of the pseudomain, but HackC, for its part, doesn't insert one there. We should be relying on the location following the .main anyway.

Reviewed By: markw65

Differential Revision: D6393664

fbshipit-source-id: eb22705294450a61f54b6dd988a14bb4a13d50e3