Functorising full_fidelity_parser_errors
Summary:
Following ericlippert's stack of functorisation diffs for the Full Fidelity Parser, we now want to functorise the lowerer (`Full_fidelity_ast`) as well. This depends on functorising `Full_fidelity_parser_errors`, which is done in this diff.

It is more involved than one would ideally expect, because it exposed a lot of "API creep." It demonstrates the need for a more throught-out design of primitive concepts (like recognising `synthesize` is a reasonable - general - primitive for `Token`s to implement).

Reviewed By: gregory-nisbet

Differential Revision: D6397564

fbshipit-source-id: bef64776ae35e153c40886f561a65d4214ca3c03