Compact contiguous blocks of .srcloc directives during HHAS linearization step
Summary: We're a bit simpleminded about the way we emit .srcloc directives (roughly every time we recurse down the AST), and end with redundant blocks of them in the HHAS. This diff compacts each contiguous block by dropping every .srcloc directive but the last. We also add a "fake" (1, 1) position to synthesized main()'s, as HHVM does when source mappings are turned on.

Reviewed By: nbenton

Differential Revision: D6388218

fbshipit-source-id: 075aeb631c907432e57a75e68b5d8958dacc0fd8