Clean up suspend event hooks
Summary:
Split onFunctionSuspendE into:
- onFunctionSuspendCreateCont (initial suspend of generators and async generators)
- onFunctionSuspendAwaitEF (suspend of eagerly executed async functions)

Split onFunctionSuspendR into:
- onFunctionSuspendAwaitR (suspend of async functions and async generators at Await opcode, previously resumed also at Await opcode)
-- currently also handles eagerly executed async generators resumed at Yield opcode (FIXME)
- onFunctionSuspendYield (suspend of generators and async generators at Yield opcode)

Makes it easier to understand event hooks and introduce a separate
onFunctionSuspendAwaitEG event hook that will deal with eagerly executed
async generators. The functionality of this event hook needs to be a mix of
previous onFunctionSuspend{E,R} hooks, as it handles eager execution on a
resumed frame and it needs to know the newly constructed AGWH, which it
currently doesn't have as it's executed before its construction.

A pre-requisite for:
- JIT support for async generators
- ResumableWaitHandle event hook support for async generators

Reviewed By: mxw

Differential Revision: D6340485

fbshipit-source-id: adf909a15842e879027ca7debfcba2490f8d3f76