enforce that XHP children are of type ?XHPChild
Summary:
Currently, we do not typecheck the body (children) expressions of an XHP literal at all. `XHPChild` is still a pretty course type with a lot of ways to circumvent it (and it doesn't yet include `Stringish`), but this will smaller fix will let us avoid blatantly incorrect things like Awaitables and other non-XHP-able things inside of XHP child expressions.

Until we improve `XHPChild`, this will check `?XHPChild`, `?Stringish`, and `?Traversable`, with the additional cases cleaned up as `XHPChild` is improved.

Reviewed By: dlreeves

Differential Revision: D6414678

fbshipit-source-id: e322810535bc0656f9c9b817a8a3eb4304ae5a35