make Stringish an XHPChild
Summary: We want to make XHPChild a more useful marker so we can type the existing XHP `appendChild` call (and relevant helpers) to accept `XHPChild` instead of having it untyped as it is today. XHP accepts Stringish (that was a large part of the reason it was created) but HHVM doesn't know that; there are a lot of Stringish floating around, but typing the children expressions will let us make XHP much safer.

Reviewed By: fredemmott

Differential Revision: D6416277

fbshipit-source-id: 684dbf1107e14b834d561dcdbbd2930169d4f1b5