Reject disposable types that extend non-disposable classes
Summary:
The escape analysis that we apply to `$this` for disposable types should be applied also
to traits that state `require implements` or `require extends` on disposable types.
Therefore we treat such traits as disposable.

In implementing this, the determination of what is a disposable type is now done in the decl
phase. This avoids full subtype tests, instead adding a new boolean to the class information.

Reviewed By: KendallHopkins

Differential Revision: D6424143

fbshipit-source-id: 6109f767feec1f24beea6086f95aa0dc1d8e26f3