Simple, brutal treatment of FPassL
Summary:
FPassL on unnamed locals is generally dangerous because of
the possibility of creating untracked aliases. We already have patterns
for some safe stylized uses of it, but this diff takes the simple
approach of forgetting any information about the local after doing an
FPassL, which seems to suffice for the remaining situations that arise
in practice.

Reviewed By: CatherineGasnier

Differential Revision: D6358985

fbshipit-source-id: d0eea5814ae30976492c6402ae9e44fa5c91de6e