Add LockFreePtrWrapper and use it for unit lineinfo
Summary: Rather than spending time in the steady state looking up unit line tables in a tbb, it would be nice to just keep a pointer to the line info on the unit. Since these tables will generally be populated early, and then be read-only for most of their existence, I wanted to optimize the read path - no locking, or writes to memory thats shared between threads. LockFreePointerWrapper is the solution I came up with. It allows readers to read a pointer (or class that contains a pointer) without any locking, and writers use the low bits of the pointer itself, along with a futex as a lock.

Reviewed By: bmaurer

Differential Revision: D6350428

fbshipit-source-id: e32128c4c312220c42c1e879b5cffa54b999a62d