Cache unit line tables
Summary:
I added some logging and found that we called loadLineTable 200000
times, for just 6000 distinct units. Adding a simple cache dropped
that to 20000. Then I noticed that more than half the time, we had two
or more entries in m_lineMap for the same line. Restructuring that to
hold a pc-range reduced the number of calls to loadLineTable to
100000, and the number of cache misses to around 15000.

Reviewed By: swtaarrs

Differential Revision: D6362818

fbshipit-source-id: 2248ff6651cecdfa379f0b8cc87837da0775aa8b