Fix subscription flushing
Summary:
In an earlier diff we changed ext_watchman to always use a new socket for every Watchman operation. Unfortunately this doesn't play nice with Watchman's flush-subscription feature which requires the flush command to be sent over the same socket used for the subscription.

We still want the property of not reusing connections, so we add a feature to `HH\watchman_sync_sub()` to do the flush automatically and this will use the correct socket.

Reviewed By: markw65

Differential Revision: D6388382

fbshipit-source-id: f4354cdf29c54fbc07063a6f9550076f82264dbd