Informant prefetches saved state before inducing restart. Server consumes cache for target handle
Summary:
Fetching the saved state package can fail for a variety of reasons (mostly
commonly due to bad certs on the machine). If it will fail, we don't
want to attempt an informant-reduced restart. So we push the fetching
earlier in the pipeline into the Informant.

The server will see the informant-prefetched package in the cache and use it.

Reviewed By: dabek

Differential Revision: D6442409

fbshipit-source-id: a24e22d0399839d1c649f1db03d098d6a6cc8ae6