Make exceptions handle cyclic previous pointers.
Summary:
Rethrowing an exception that is currently being unwound in a finally block results in a cyclic previous pointer in the Exception.  We were failing to print out these exceptions; either OOMing the request or crashing the VM in certain circumstances (T23794970).
I checked out PHP src, and they avoid cyclicly printing the `__toString` as well.

For T23794970 I was unable to identify where in WWW the exception was being rethrown, but I assume this is the issue, and not a bug in the unwinder somewhere.

Reviewed By: markw65

Differential Revision: D6409128

fbshipit-source-id: 079f974da7ab93f7f0ef78aa55038ca9af3f9785