Declaration and naming for reactive types
Summary:
This adds reactivity to the declaration step, as well as syntax for reactive function types:
Any typehint of the form
```
Rx<function(params):ret>
```
will be interpreted as a fun_ty with the params and return as before, with is_reactive set to true.

We don't do anything with the reactivity information yet: that's the next diff.

Reviewed By: KendallHopkins

Differential Revision: D6317378

fbshipit-source-id: e069c8e5b6a278ee381a2ec01d2e2cec6731100d