Namespace issues with built-in types
Summary:
Initial work on sorting out some of the issues we have with "auto" importing of HH types
into namespaces. For code such as
```
namespace MyNamespace {
  class Map { ... }
  function foo(Map $x) { ... }
}
```
we now treat `Map` specially. Rather than auto-importing it into `HH\Map` we treat it in the `namespaces.ml` elaboration phase as if
there were a `use type MyNamespace\Map as Map` declaration.

There is still a lot of duplication of functionality between `namespaces.ml`, `Hhbc_id.ml` and `Hhbc_alias.ml` which will be simplified
in future diffs.

Reviewed By: hubyrod

Differential Revision: D6447045

fbshipit-source-id: e1bfa05cddf3ce5d9009b02f24abf24d090229f3