Abort+retry if server still hasn't sent "hello" after 200secs
Summary:
Normally you call ClientConnect.connect, and that gives an ic/oc channel pair, and you wait for the server to send you a "hello" message. That should come as soon as the server has finished typechecking/initializing.

One user observed that the Hack server was restarted, connected, but still hadn't sent a "hello" message after 731 seconds (at which point he gave up).

In general, if saved-states are working then we expect initialization in under 100 seconds. If not, we expect initialization in 400seconds.

This diff adds a last-resort bailout as a bandaid: if we still haven't received the "hello" after 200 seconds, then we just close the ic/oc pair, and call ClientConnect.connect all over again. That should avoid users getting stuck with nowhere to go.

Reviewed By: arxanas

Differential Revision: D6454463

fbshipit-source-id: 4e4c5f6a042a883e858ecdb3974547872d0047a0