ClientLsp/rage gathers logfiles even without server
Summary:
Currently when the LSP client sends a telemetry/rage LSP request (e.g. because the user clicked a "file-a-bug" button) then clientLsp asks the hh_server to provide logfiles.

But even if the server is down (e.g. it's still initializing, or it got disconnected) we'd still like to gather logs\!

This diff changes it so clientLsp is able to gather server and monitor logs even in the absence of a working persistent connection.

Reviewed By: arxanas

Differential Revision: D6460189

fbshipit-source-id: d832a9b35a30de2ecfbe3fdd3dd9bc66e61ed3d7