Fix Stringish interfaces
Summary:
InterfaceMap::Builder::add doesn't do the right thing if you add an
interface twice. In dbg builds, it asserts, but in release builds,
looking up the interface will get the wrong index.

D6416277 auto adds XHPChild as an interface for classes that define a
toString method, but doesn't check to see if XHPChild has already been
added.

This adds the check, and also adds a more convenient "contains" method
to IndexedStringMap::Builder.

Reviewed By: ngavalas, alexeyt

Differential Revision: D6470934

fbshipit-source-id: ea430a982c9ecf2a597f12f1350d8cd42040f8a1