Ensure top-level this goes at end of .declvars
Summary: Yuck, but this can be captured by includes/requires inside class in a way analagous to the way they are inside lambdas. See how mod.inc is required in fbcode/hphp/test/quick/reqonce.php. So we imitate the logic of  https://fburl.com/w21h5g56 and put this at the end of the list of .declvars in a function.

Reviewed By: vladima

Differential Revision: D6461173

fbshipit-source-id: 0bb1e703f42d4aee691c6d47da565811d613dffc