Trigger RetranslateAll based on server uptime
Summary:
This diff extends the trigger mechanism for RetranslateAll to also take the
server uptime into account.  The previous mechanism, based on the number of
requests, is also preserved.  RetranslateAll, when enabled, will now be
triggered whenever either of these conditions is met.  The default value for the
number of requests to trigger RetranslateAll is bumped very high, so that, in
practice, the server uptime is going to be the effective trigger.

Reviewed By: paulbiss

Differential Revision: D6468253

fbshipit-source-id: 2df7b3bdc60fb06e5882d2a942f235271921b5e9