Remove unnecessary branch in mallocSmallSizeSlow
Summary:
When we run out of freelist objects, we attempt to find
a larger object to split. The search for a larger object
is guaranteed to start at a large size class than the
empty freelist, and so when we split it, we're guaranteed
to have a nonzero remainder after the first object in
the split block.

Reviewed By: alexeyt

Differential Revision: D6473863

fbshipit-source-id: 8dcb7b4c03f1b8e0c82245cdc6b2f0db3f1695a7