Better weaken_decrefs, and irrelevant_inst
Summary:
If a DecRef is followed by a use of the SSATmp, we should generally be able to assume that the DecRef doesn't cause the refcount to hit zero. In weaken_decrefs, we didn't do that, because there are a few special cases where a subsequent use of an SSATmp is allowed - in particular PureStores are allowed to write the value of the SSATmp to memory (in that case, it must be a dead store, but we may not be able to prove its a dead store). This codifies the special cases, and allows weaken_decrefs to do a lot more.

Also there are a lot of instructions that return an empty GeneralEffects because although they can't manipulate refcounts, they may be able to observe them. The remove_trivial_incdecs pass can ignore these, since it only needs to know that the refcounts aren't altered between the Inc and Dec (or Dec and Inc).

Reviewed By: ricklavoie

Differential Revision: D6467508

fbshipit-source-id: 55339366cda830fdf8ad29fa084f3b40c5681665