Return the smallest node at the given position in ServerInferType
Summary:
Previously, we simply returned the type of the final expression visited in a preorder traversal of the tree (modulo special handling for Call expressions). This is only correct if a node with a position P always contains every other node with a position contained by P. That isn't actually the case in an AnnotatedAST--when we desugar `invariant`, we build an AST that violates this property.

This change works around the problem by selecting the smallest node instead of the last one visited.

Reviewed By: ljw1004

Differential Revision: D6457018

fbshipit-source-id: 049d50499098755a462f3d7cdb743f0784e51487