Add Typing_dep.Dep.AllMembers
Summary:
Currently whenever a new elemnt is added to an enum it requires checking all files that use any element of the enum. This means even if a file refers to a single element of an enum it needs to be rechecked  when a new element is added. We do this as a simple way of handling exhaustiveness checks for enums.

This diffs adds a new dependency in Typing_deps.Dep.AllMembers. This specifies that a file has a dependency on all members of a class. If any member changes, a new member is added or deleted then the file is rechecked. We add this dependency when doing exhaustiveness checks.

Reviewed By: jamesjwu

Differential Revision: D6469069

fbshipit-source-id: cf3457fce94b5fafb0b7a136e8f7beeea43da1dd