lsp: catch error returned from a dialog
Summary:
Logs show the commonest error experienced by clientLsp is when a response comes back from a showMessageRequest, but the response isn't the expected MessageActionItem object.

I'm not sure what exactly triggers this. Maybe it's when the user dismisses the dialog and the LSP client sends back an empty response? Or maybe it's when the LSP server requests the showMessageRequest to be cancelled, and the server responds back with a CancelledError as expected?

In any case, I'm just going to catch this kind of issue because it's not really an error and shouldn't show up in the logs as such.

(Note: the behavior of `Hh_json_helpers.get_string_val` is to throw an exception if it's being asked to operate on something that isn't a JSON_Object).

Reviewed By: arxanas

Differential Revision: D6471582

fbshipit-source-id: 44b16acdb25380359641b5846e8efec2c9e70e00