Mark functions, methods interceptable in HHAS iff among DynamicInvokeFunctions
Summary: Functions and methods can be explicitly designated as interceptable with the DynamicInvokeFunctions parameter. With this diff, we read this parameter from the inherited HHVM config and mark functions as "interceptable" in the generated HHAS. (We incidentally make it possible to parse and print this function attribute in HHAS.) Depends on D6290975.

Reviewed By: paulbiss

Differential Revision: D6433699

fbshipit-source-id: fe304a8e2c9701430b8755a95aa9e5aae65665f0