Typing step for is-expression, part 2
Summary: This implements the type refinement portion of `is` expressions (for the types that we have written the validation logic for). Future diffs will continue to add support for the remaining types. If a typehint cannot be used in an `is` expression, we do not refine the type of the target variable, EXCEPT in the case of classes with generics. This allows unsafe behavior, but isolates the typechecker error in one part of the code, instead of scattering it around everywhere with the `Tv#n`-style generics.

Reviewed By: dlreeves

Differential Revision: D6432217

fbshipit-source-id: bf449e4227dc13ffc41f820f30ed85850a1c837c