Typing step for is-expression, part 2.5
Summary:
This updates the hint validator to be tri-state. The previous implementation of valid/invalid wasn't descriptive enough, because there are actually three situations we care about in the typechecker:
- Invalid: completely unsupported, like generics and functions
- Partially valid: Raises a typechecker error, but still refines the type (classes with generics)
- Valid: Refines the type

This also fixes a bug in the class type validator in which we didn't validate the type parameters, meaning we would've allowed something like `$foo is MyClass<T>`. Now, since invalid types always take priority, that expression should be invalid.

Reviewed By: dlreeves

Differential Revision: D6460968

fbshipit-source-id: f615580ad9380ea08d5eeb31f9725f08e7366469