Set finally label only if it is used
Summary:
Label in finally was set unconditionally which lead to a extra EntryNop instruction being added in case of body of try block was empty. This extra instruction confused semdiff since HackC didn't emit it.
This diff sets the label only if it was used before - similar idiom was used in other places of the emitter.

Reviewed By: markw65

Differential Revision: D6487124

fbshipit-source-id: 0b4bd6ed578d8bf7df5e36706debd869fb2f5c81