Fixed `do_definition` to prevent jumping to base class
Summary: Cmd+Click (GoToDef) will no longer jump to a base class if that's where the most-derived constructor is defined. We do this by first filtering the results in `do_definition` down to those whose definition occurs in the class that is being clicked on. We then check to see if any of the remaining results are methods, so that we can jump to a constructor if it is defined. If no method exists, we will jump to the beginning of the class definition.

Reviewed By: ljw1004

Differential Revision: D6291848

fbshipit-source-id: 205b3e5bef20a4340ceb068ec1bd2f1ef7abe833