Deal with some systemlib/script destructors
Summary:
Memcache::__destruct() calls $this->close(), which calls
memcached_quit(). MemcacheData's native destructor calls memcached_free(),
which already calls memcached_quit() when needed [1]. As far as I can tell,
__destruct() wasn't doing anything useful. Empty it out and add attribute
__OptionalDestruct, in case users are subclassing it somewhere and explicitly
calling it.

finfo::__destruct() is similar: it calls finfo_close($this->resource), which
calls FileinfoResource::close(). The latter is already called by
FileinfoResource's native destructor and sweep functions, so again, the
userspace destructor wasn't doing anything useful.

Intentionally not handled in this diff:

UConverter::__destruct() can reenter user code (see T4017519) and I see no
signs of it being used outside of tests, so punt on it for now. If anyone tries
to create one in an OBRC build, it will fatal.

OAuth is part of EZC, and therefore doesn't matter in the OBRC world.

XhprofFrame is intentionally an RAII-style object. It's currently unused in our
codebase, and if we ever need a replacement, we can build a disposable version.

[1]: http://docs.libmemcached.org/memcached_quit.html

Reviewed By: alexeyt

Differential Revision: D6492093

fbshipit-source-id: 01acf18cf31f43af25349a718e5ca7283399600f