Proper fix: retry if server "hello" fails to arrive
Summary:
We've observed some cases where the "Hello" from the server never arrives.

Previously we put in a quick band-aid fix which, if more than 200s had elapsed since we first established the connection, then it called "do_lost_server" which resets the timer and makes the "waiting for hh_server" dialog to reappear.

This diff has a nicer fix. Now if it's been more than just 10s waiting for a "Hello" (maybe because the server is initializing, or type-checking, or because the "Hello" just got lost), then it closes the connection and creates a new one. But it does so without reseting the timer or the UI.

Reviewed By: arxanas

Differential Revision: D6495691

fbshipit-source-id: 57f33d76b1c829aceec43c83bc64d6d6fe05a944