followup to make Stringish an XHPChild
Summary: This completes the process of making Stringish an XHPChild by informing Hack about it and cleaning up the relevant union case in `Typing_xhp`. This is safe to land now since no one is relying on the behavior of `instanceof`, but I can also wait to land this until the HHVM release ships with my runtime changes.

Reviewed By: dlreeves

Differential Revision: D6428850

fbshipit-source-id: ab2e09bd069f41b99ecaaa180c59e83bba7866b1