Implement reactive lambda functions
Summary:
This diff is an attempt to implement the correct typechecking behavior for lambdas in Hack.

This is the behavior given:
- Lambdas are statically checked on declaration. If they do not exhibit any actions that prevent them from being reactive, they are typed as reactive.
- Since reactive functions are subtypes of regular ones, lambdas in regular code have no change in behavior
- Lambdas that access global state or assign to objects without mutating them are typed as nonreactive.

Lambdas in hack are a pain to typecheck, as they are implemented with... well.. lambdas. The approach I went with was to set a flag that runs through typechecking code and checks if the block contains reactive behavior, setting a flag if it does.

Reviewed By: KendallHopkins

Differential Revision: D6454108

fbshipit-source-id: 67989855cc3786d9e151bfa1f3b31b8209dba5e3