Use Dollar instead of Lvarvar
Summary:
To better support syntax such as `${x}` and `$$x` and `${$x}` we now have the `Dollar` node,
so `Lvarvar` is redundant. Instead, we now represent `Lvarvar(n,id)` as `Dollar(...(Dollar(Lvar id)))` with `n` occurrences of the `Dollar` node.

For now, we leave `Lvarvar` in the AST and NAST. It will be removed in a subsequent diff.

Reviewed By: hubyrod, vladima

Differential Revision: D6519616

fbshipit-source-id: c3ab47d4a27606e42499fb490bd575b353b97469