Retain bracedness in lowering
Summary: In an earlier diff (D6079143), we peeled off braced expressions from their contained expressions, to make the XHP positions correct. Codegen requires the information that the expression was braced, though. This puts the `BracedExpr` constructor in the result again, *except* if it simply encloses a string or an Lvar, in which case we elide it to simplify code gen.

Reviewed By: oulgen

Differential Revision: D6323257

fbshipit-source-id: 5a8e776ccff88e06471ffbf0bdfca5eec4457578