compute estimated stack depth for inline HHAS
Summary:
inline HHAS is treated as expression so if after running it evaluation stack is left intact - emitter should push Null. This diff add ability to compute estimated stack depth for a provided instruction sequence (it does not try to verify its correctness).

Support for inline HHAS should also allow to replace the entire method bodies including values for .declvars, .numiters etc that a currently not represented as instructions - will be added in subsequent diffs

Reviewed By: hubyrod

Differential Revision: D6532516

fbshipit-source-id: 9f0ecfa1fccfb0efc6fa9bfd79c325cca88a7a26