Pretty-print File_pos.t using column instead of byte offset
Summary:
Today, File_pos.t is pretty-printed as `L<line>:<cnum>`, where `<line>` is a 1-based line number, and `<cnum>` is a 0-based byte offset in the file (Consequently, Pos.t is pretty-printed `[L<line>:<cnum>-L<line>:<cnum>]`). Since `cnum` isn't very useful to a human reader, this diff changes it to a 1-based column number. Pos.t is a half-open (inclusive-exclusive) interval, so the interval we print after this change is too.

This change also removes the L, because I'm not sure why it needed to be there.

Reviewed By: jamesjwu

Differential Revision: D6516298

fbshipit-source-id: 58af8ecac8ee09ab7752b00e6407a97edd98ac97