Stop filtering positions from TAST test cases
Summary:
Presumably these were filtered to keep test case output from being too noisy, but positional information is critical and should be tested. I'm not particularly concerned about the fact that all the positions change when the source file is edited--we don't edit these test cases often.

Since the position filtering is just a regex replace, it doesn't seem necessary to include it as an option in hh_single_type_check, so in addition to turning it off this change just removes it altogether.

Reviewed By: arxanas

Differential Revision: D6516299

fbshipit-source-id: a1f1d876e33638d1f53f965765aea7ecf4654d65