Rewrite Diagnostic_subscription to use new error map
Summary: Treat it as a ground up rewrite of this module and don't pay attention to what was there before. The previous approach was fundamentally broken, because it was based on the idea that "if we want to get errors in A, we need to recheck A". This is not really true, as there can be other files contributing to errors in A, and we would sometimes loose them in IDE (I suspect this is one of the sources of "I see the errors in command line, but not in editor" bug reports).

Reviewed By: alexchow

Differential Revision: D6393208

fbshipit-source-id: f605df9943219625b64f98e52bceb02856c14731