Do not recheck files only because they had errors before
Summary:
In incremental mode, we are constantly rechecking all files with errors. For example, if you change file A which generates errors e1...e100 in files B1.php...B100.php, and then proceed to fix the error in B52, we'll still recheck all 100 files, even though only one of the errors have changed. I believe that there are only three reasons why we do it:

- simplicity - in batch mode it doesn't hurt as much, but if we want to support IDE scenarios it gets bad
- e1...e100 is a list, and we didn't knew which part of the list we need to replace. After {D6393159} we have a map that allows us to change the parts that we want so it's not an issue anymore
- positions: we try to not do the work for insignificant changes: let's say that ERRORS(C) depends on DECL(B) which depends on DECL(A) which depends on AST(A). If we make an insignificant change to A (say, adding an empty line), we will notice that AST(A) have changed, but DECL(A/B) didn't (modulo positions), so we will not proceed to recompute ERRORS(C) (or any other errors depending on DECL(A/B), because it's impossible that any errors were added or removed). But ERROR(C) can actually contain some positions from AST(A), so we always just recompute all errors to keep them up to date.

In this diff I change it to recompute only errors that contain positions that COULD have changed - i.e positions pointing to files that actually changed on disk. You can think about it as having a direct edge between ERRORS(C) and AST(A) in dependency graph.

This code is not very well understood and there is some risk that there is another reason why we keep rechecking those things that I don't know about and something will break - fingers crossed!

Reviewed By: alexchow

Differential Revision: D6393223

fbshipit-source-id: 6a4153b9df6458be4803833719f0c21099e04cbe